dai 20210129 dai
第一版测试

+++ pro/src/main/java/com/zx/pro/controller/OrderInfoController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpSession;
 import java.util.HashMap;
 import java.util.List;
 
@@ -28,8 +29,13 @@ public class OrderInfoController {
     private IOrderInfoService iOrderInfoService;
 
     @RequestMapping("/post_list")
-    public Object postList() {
+    public Object postList(HttpSession session) {
         try {
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isSelect("订单汇总")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             List<OrderInfoItem> list = iOrderInfoService.getList();
 
             if (StringUtils.isNotNull(list)) {
@@ -45,8 +51,13 @@ public class OrderInfoController {
     }
 
     @RequestMapping("/select_list")
-    public ResultInfo postList(@RequestBody HashMap map) {
+    public ResultInfo postList(HttpSession session,@RequestBody HashMap map) {
         try {
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isSelect("订单汇总")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             String projectName = map.get("projectName").toString();
             String orderId = map.get("orderId").toString();
             String startDate = map.get("startDate").toString();
@@ -68,8 +79,13 @@ public class OrderInfoController {
     }
 
     @RequestMapping("/post_state_list")
-    public ResultInfo postStateList() {
+    public ResultInfo postStateList(HttpSession session) {
         try {
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isSelect("订单状态")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             List<OrderState> list = iOrderInfoService.getOrderStateList("");
 
             if (StringUtils.isNotNull(list)) {
@@ -85,8 +101,13 @@ public class OrderInfoController {
     }
 
     @RequestMapping("/select_state_list")
-    public ResultInfo postStateList(@RequestBody HashMap map) {
+    public ResultInfo postStateList(HttpSession session,@RequestBody HashMap map) {
         try {
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isSelect("订单状态")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             String orderId = map.get("orderId").toString();
 
             List<OrderState> list = iOrderInfoService.getOrderStateList(orderId);
@@ -105,8 +126,13 @@ public class OrderInfoController {
     }
 
     @RequestMapping("/update")
-    public ResultInfo update(@RequestBody String orderInfoItemJson){
+    public ResultInfo update(HttpSession session,@RequestBody String orderInfoItemJson){
         try{
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isUpdate("订单汇总")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             OrderInfoItem orderInfoItem = DecodeUtil.decodeToJson(orderInfoItemJson,
                     OrderInfoItem.class,"createTime");
 
@@ -125,10 +151,15 @@ public class OrderInfoController {
     }
 
     @RequestMapping("/delete")
-    public ResultInfo delete(@RequestBody HashMap map){
+    public ResultInfo delete(HttpSession session,@RequestBody HashMap map){
         GsonUtil gsonUtil = new GsonUtil(GsonUtil.toJson(map));
 
         try{
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isDelete("订单汇总")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             List<Integer> idList = GsonUtil.toList(gsonUtil.get("idList"),Integer.class);
 
             if(iOrderInfoService.delete(idList)){
@@ -146,10 +177,15 @@ public class OrderInfoController {
 
     @RequestMapping("/add")
     @Transactional
-    public ResultInfo add(@RequestBody HashMap map){
+    public ResultInfo add(HttpSession session,@RequestBody HashMap map){
         GsonUtil gsonUtil = new GsonUtil(GsonUtil.toJson(map));
 
         try{
+            PowerUtil powerUtil = PowerUtil.getPowerUtil(session);
+            if(!powerUtil.isAdd("订单汇总")){
+                return ResultInfo.error(401,"无权限");
+            }
+
             OrderInfo orderInfo = GsonUtil.toEntity(gsonUtil.get("orderInfo"), OrderInfo.class);
             List<HashMap> list = GsonUtil.toList(gsonUtil.get("productInfoList"),HashMap.class);
